pyproject.toml
[project]
name = "medical-ocr-pipeline"
version = "0.1.0"
description = "FastAPI-based OCR → LLM → PDF Filling Pipeline for medical forms"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI & Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    # Validation & Config
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # PDF Processing
    "pypdf>=5.1.0",
    "pypdfium2>=4.30.0",
    "pdf2image>=1.17.0",
    "Pillow>=11.0.0",
    "pillow-heif>=0.21.0",
    "pymupdf>=1.25.0",
    # HTTP Client
    "httpx>=0.28.0",
    # LLM SDKs
    "openai>=1.58.0",
    "google-generativeai>=0.8.0",
    "mistralai>=1.2.0",
    "reportlab>=4.4.7",
    "markdown>=3.10",
    "fpdf2>=2.8.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
